Move test_unload_entry
