+++ tests/components/here_weather/test_config_flow.py
@@ -11,7 +11,6 @@ from homeassistant.const import (
     CONF_LONGITUDE,
     CONF_NAME,
     CONF_SCAN_INTERVAL,
-    EVENT_HOMEASSISTANT_START,
 )
 from homeassistant.core import HomeAssistant
 
@@ -157,29 +156,3 @@ async def test_options(hass):
         await hass.async_block_till_done()
         assert result["type"] == "create_entry"
         assert result["data"][CONF_SCAN_INTERVAL] == 10
-
-
-async def test_unload_entry(hass):
-    """Test unloading a config entry removes all entities."""
-    with patch(
-        "herepy.DestinationWeatherApi.weather_for_coordinates",
-        side_effect=mock_weather_for_coordinates,
-    ):
-        entry = MockConfigEntry(
-            domain=DOMAIN,
-            data={
-                CONF_API_KEY: "test",
-                CONF_NAME: DOMAIN,
-                CONF_LATITUDE: "40.79962",
-                CONF_LONGITUDE: "-73.970314",
-            },
-        )
-        entry.add_to_hass(hass)
-        await hass.config_entries.async_setup(entry.entry_id)
-        await hass.async_block_till_done()
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
-        await hass.async_block_till_done()
-        assert hass.data[DOMAIN]
-        assert await hass.config_entries.async_unload(entry.entry_id)
-        await hass.async_block_till_done()
-        assert not hass.data[DOMAIN]
