+++ tests/components/here_weather/test_init.py
@@ -1,0 +1,34 @@
+"""Tests for the here_weather integration."""
+
+from unittest.mock import patch
+
+from homeassistant.components.here_weather.const import DOMAIN
+from homeassistant.const import CONF_API_KEY, CONF_LATITUDE, CONF_LONGITUDE, CONF_NAME
+
+from . import mock_weather_for_coordinates
+
+from tests.common import MockConfigEntry
+
+
+async def test_unload_entry(hass):
+    """Test unloading a config entry removes all entities."""
+    with patch(
+        "herepy.DestinationWeatherApi.weather_for_coordinates",
+        side_effect=mock_weather_for_coordinates,
+    ):
+        entry = MockConfigEntry(
+            domain=DOMAIN,
+            data={
+                CONF_API_KEY: "test",
+                CONF_NAME: DOMAIN,
+                CONF_LATITUDE: "40.79962",
+                CONF_LONGITUDE: "-73.970314",
+            },
+        )
+        entry.add_to_hass(hass)
+        await hass.config_entries.async_setup(entry.entry_id)
+        await hass.async_block_till_done()
+        assert hass.data[DOMAIN]
+        assert await hass.config_entries.async_unload(entry.entry_id)
+        await hass.async_block_till_done()
+        assert not hass.data[DOMAIN]
